Added some src files
These files still need to be refactored by me over the weekend but I at least started adding them in.

I need to modify the motor wrapper motor_pins enum to be a struct so I can make one per motor, and it makes changing which pins we use easier.